lambda/discovery/requirements.txt
pyiceberg[s3fs,glue]
boto3
python-dotenv
pyarrow
pandas
